// src/api/api.ts
const BASE_PATH = "https://localhost/api".replace(/\/+$/, "");

/**
 *
 * @export
 */
export const COLLECTION_FORMATS = {
    csv: ",",
    ssv: " ",
    tsv: "\t",
    pipes: "|",
};

/**
 *
 * @export
 * @interface FetchAPI
 */
export interface FetchAPI {
    (url: string, init?: any): Promise<Response>;
}

/**
 *  
 * @export
 * @interface FetchArgs
 */
export interface FetchArgs {
    url: string;
    options: any;
}

/**
 * 
 * @export
 * @class BaseAPI
 */
export class BaseAPI {
    protected configuration: Configuration;

    constructor(configuration?: Configuration, protected basePath: string = BASE_PATH, protected fetch: FetchAPI = portableFetch) {
        if (configuration) {
            this.configuration = configuration;
            this.basePath = configuration.basePath || this.basePath;
        }
    }
};

/**
 * 
 * @export
 * @class RequiredError
 * @extends {Error}
 */
export class RequiredError extends Error {
    name: "RequiredError"
    constructor(public field: string, msg?: string) {
        super(msg);
    }
}

/**
 * 
 * @export
 * @interface ClassCount
 */
export interface ClassCount {
    /**
     * 
     * @type {string}
     * @memberof ClassCount
     */
    classname?: string;
    /**
     * 
     * @type {number}
     * @memberof ClassCount
     */
    instancecount?: number;
}

/**
 * 
 * @export
 * @interface Dataset
 */
export interface Dataset {
    /**
     * 
     * @type {string}
     * @memberof Dataset
     */
    datasetId?: string;
    /**
     * 
     * @type {string}
     * @memberof Dataset
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof Dataset
     */
    numclass?: number;
    /**
     * 
     * @type {number}
     * @memberof Dataset
     */
    numfolds?: number;
    /**
     * 
     * @type {Array&lt;string&gt;}
     * @memberof Dataset
     */
    classes?: Array<string>;
    /**
     * 
     * @type {Array&lt;Fold&gt;}
     * @memberof Dataset
     */
    folds?: Array<Fold>;
}

/**
 * 
 * @export
 * @interface EpochData
 */
export interface EpochData {
    /**
     * 
     * @type {number}
     * @memberof EpochData
     */
    epochId?: number;
    /**
     * 
     * @type {Array&lt;number&gt;}
     * @memberof EpochData
     */
    confmat?: Array<number>;
}

/**
 * 
 * @export
 * @interface Fold
 */
export interface Fold {
    /**
     * 
     * @type {string}
     * @memberof Fold
     */
    foldId?: string;
    /**
     * 
     * @type {string}
     * @memberof Fold
     */
    description?: string;
    /**
     * #/definitions/Dataset/properties/datasetId
     * @type {string}
     * @memberof Fold
     */
    dataset?: string;
    /**
     * 
     * @type {number}
     * @memberof Fold
     */
    numinstances?: number;
    /**
     * 
     * @type {Array&lt;ClassCount&gt;}
     * @memberof Fold
     */
    classcounts?: Array<ClassCount>;
}

/**
 * 
 * @export
 * @interface FoldLog
 */
export interface FoldLog {
    /**
     * 
     * @type {string}
     * @memberof FoldLog
     */
    foldlogId?: string;
    /**
     * 
     * @type {string}
     * @memberof FoldLog
     */
    description?: string;
    /**
     * #/definitions/Run/properties/runId
     * @type {string}
     * @memberof FoldLog
     */
    runId?: string;
    /**
     * #/definitions/Fold/properties/foldId
     * @type {string}
     * @memberof FoldLog
     */
    foldId?: string;
    /**
     * 
     * @type {number}
     * @memberof FoldLog
     */
    numepochs?: number;
}

/**
 * 
 * @export
 * @interface FoldLogData
 */
export interface FoldLogData {
    /**
     * 
     * @type {string}
     * @memberof FoldLogData
     */
    foldlogId?: string;
    /**
     * 
     * @type {number}
     * @memberof FoldLogData
     */
    numepochs?: number;
    /**
     * 
     * @type {Array&lt;EpochData&gt;}
     * @memberof FoldLogData
     */
    epochdata?: Array<EpochData>;
}

/**
 * 
 * @export
 * @interface HyperParam
 */
export interface HyperParam {
    /**
     * 
     * @type {string}
     * @memberof HyperParam
     */
    arch?: string;
    /**
     * 
     * @type {string}
     * @memberof HyperParam
     */
    optim?: string;
    /**
     * 
     * @type {number}
     * @memberof HyperParam
     */
    bs?: number;
    /**
     * 
     * @type {Array&lt;number&gt;}
     * @memberof HyperParam
     */
    lr?: Array<number>;
    /**
     * 
     * @type {Array&lt;number&gt;}
     * @memberof HyperParam
     */
    m?: Array<number>;
}

/**
 * 
 * @export
 * @interface Run
 */
export interface Run {
    /**
     * 
     * @type {string}
     * @memberof Run
     */
    runId?: string;
    /**
     * #/definitions/Fold/properties/foldId
     * @type {string}
     * @memberof Run
     */
    trainfoldId?: string;
    /**
     * 
     * @type {HyperParam}
     * @memberof Run
     */
    hyperparam?: HyperParam;
    /**
     * 
     * @type {Array&lt;FoldLog&gt;}
     * @memberof Run
     */
    foldlogs?: Array<FoldLog>;
}
// ...
export class DatasetApi extends BaseAPI {
    /**
     * Returns a single dataset
     * @summary Get dataset by ID
     * @param {} datasetId ID of dataset to return
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DatasetApi
     */
    public getDatasetById(datasetId: string, options?: any) {
        return DatasetApiFp(this.configuration).getDatasetById(datasetId, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @summary Show all available datasets
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DatasetApi
     */
    public getDatasets(options?: any) {
        return DatasetApiFp(this.configuration).getDatasets(options)(this.fetch, this.basePath);
    }

}
// ...
export class FoldlogApi extends BaseAPI {
    /**
     * Returns a single foldlog
     * @summary Get foldlog by ID
     * @param {} foldlogId ID of foldlog to return
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FoldlogApi
     */
    public getFoldLogById(foldlogId: string, options?: any) {
        return FoldlogApiFp(this.configuration).getFoldLogById(foldlogId, options)(this.fetch, this.basePath);
    }

    /**
     * Returns data for a single foldlog
     * @summary Get FoldLogData by ID
     * @param {} foldlogId ID of foldlog to return data
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FoldlogApi
     */
    public getFoldLogDataById(foldlogId: string, options?: any) {
        return FoldlogApiFp(this.configuration).getFoldLogDataById(foldlogId, options)(this.fetch, this.basePath);
    }

}
// ...
export class RunApi extends BaseAPI {
    /**
     * Returns a single run
     * @summary Get run by ID
     * @param {} runId ID of run to return
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RunApi
     */
    public getRunById(runId: string, options?: any) {
        return RunApiFp(this.configuration).getRunById(runId, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @summary Show all available runs
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof RunApi
     */
    public getRuns(options?: any) {
        return RunApiFp(this.configuration).getRuns(options)(this.fetch, this.basePath);
    }

}